clean up, making common code common
