--- chain/ethereum/geth/types.go
@@ -1,1 +1,0 @@
-package geth
