+++ chain/ethereum/wallet.go
@@ -1,4 +1,4 @@
-package foundry
+package ethereum
 
 import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
@@ -8,14 +8,16 @@ import (
 var _ ibc.Wallet = &EthereumWallet{}
 
 type EthereumWallet struct {
-	address string
+	address []byte
 	keyName string
+	mnemonic string
 }
 
-func NewWallet(keyname string, address string) ibc.Wallet {
+func NewWallet(keyname string, address []byte, mnemonic string) ibc.Wallet {
 	return &EthereumWallet{
-		address: address,
-		keyName: keyname,
+		address:  address,
+		keyName:  keyname,
+		mnemonic: mnemonic,
 	}
 }
 
@@ -25,29 +27,15 @@ func (w *EthereumWallet) KeyName() string {
 
 // Get formatted address, passing in a prefix
 func (w *EthereumWallet) FormattedAddress() string {
-	return w.address
+	return hexutil.Encode(w.address)
 }
 
 // Get mnemonic, only used for relayer wallets
 func (w *EthereumWallet) Mnemonic() string {
-	return ""
+	return w.mnemonic
 }
 
 // Get Address with chain's prefix
 func (w *EthereumWallet) Address() []byte {
-	return hexutil.MustDecode(w.address)
-}
-
-type GenesisWallets struct {
-	total uint32
-}
-
-func NewGenesisWallet() GenesisWallets {
-	return GenesisWallets{
-		total: 2, // Start with 2 at genesis, one for faucet, one for relayer
-	}
-}
-
-func (w *GenesisWallets) GetFaucetWallet(keyname string) ibc.Wallet {
-	return NewWallet(keyname, "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266")
+	return w.address
 }
